Add on hover cart box and store cart item in local Storage
