+++ src/components/auth/auth.js
@@ -4,7 +4,7 @@ import { Link } from "react-router-dom";
 import { Dropdown } from "antd";
 import { connect } from "react-redux";
 import UserMenu from "./userMenu";
-import { onAuthStateChanged,getAuth } from "firebase/auth";
+import { onAuthStateChanged, getAuth } from "firebase/auth";
 //import { auth } from "../../config/firebase-config";
 
 function Auth(props) {
@@ -13,11 +13,13 @@ function Auth(props) {
   const [userName, setUserName] = useState(null);
   const { dispatch } = props;
   const auth = getAuth();
-  
+
   const generateUserInitial = (fullName) => {
     const nameArray = fullName.split(" ");
     const firstNameInitial = nameArray[0].charAt(0).toUpperCase();
-    const lastNameInitial = nameArray[nameArray.length - 1].charAt(0).toUpperCase();
+    const lastNameInitial = nameArray[nameArray.length - 1]
+      .charAt(0)
+      .toUpperCase();
     if (nameArray.length === 1) {
       return firstNameInitial;
     }
@@ -26,10 +28,10 @@ function Auth(props) {
 
   onAuthStateChanged(auth, (currentUser) => {
     if (currentUser) {
-      console.log(currentUser)
+      console.log(currentUser);
       dispatch({ type: "USER_NAME", payload: currentUser.displayName });
     }
-  })
+  });
 
   useEffect(() => {
     console.log(props.userName);
