+++ src/components/cart/cardStyledComponent.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const CartBox = styled.div`
+  border-radius: 5px;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  height: 500px;
+  overflow-y: scroll;
+`;
+
+export { CartBox };
