+++ src/reducer/cart.js
@@ -1,5 +1,6 @@
 const initialState = {
   cart: [],
+  isCartUpdated: false
 };
 
 export default function CartReducer(state = initialState, action) {
@@ -9,6 +10,13 @@ export default function CartReducer(state = initialState, action) {
         // Again, one less level of nesting to copy
         ...state,
         cart: action.payload,
+      };
+    }
+    case "CART_UPDATED": {
+      return {
+        // Again, one less level of nesting to copy
+        ...state,
+        isCartUpdated: action.payload,
       };
     }
     default:
